Added some table validation
This now clamps the table coordinates to be within 5% and 95%, so it shouldn't go off the edge of the canvas. I had lots of issues with that when working out the drag and drop stuff, particularly for lines, but this should protect it even if the drag and drop code isn't perfect.

Capacity now needs to be at least 1.