More consistent table placement, customizable table size
Keeping a 1:1 aspect ratio on the table container keeps the tables from moving around based on the width and height of the page, so it should stay consistent.

This also adds a range slider that lets you change the visible size of the tables, for if you need to stuff a bunch of extra tables in.